refactor: app 디렉토리 생성 및 App 컴포넌트 위치 변경
