+++ front-end/src/index.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "./app/App";
 
 ReactDOM.render(
   <StrictMode>
